+++ src/pages/jihyeon/Main/ArticleJihyeon/ArticleJihyeon.scss
@@ -1,0 +1,321 @@
+.articleJihyeon {
+  width: 70vw;
+  height: 82vh;
+  margin-top: 10vh;
+  margin-bottom: 0;
+  margin-left: 3vw;
+  flex-direction: column;
+  background-color: white;
+
+  .articleProfile {
+    height: 6vh;
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+    position: relative;
+
+    .articleProfilePhoto {
+      margin-left: 3vw;
+    }
+
+    .articleProfileName {
+      margin-left: 1vw;
+    }
+
+    .articleProfileOptBtn {
+      position: absolute;
+      padding: 0;
+      right: 0;
+      margin-right: 2.5vw;
+      height: 5vh;
+      width: 5vh;
+      background-color: inherit;
+      border: 0;
+
+      img {
+        height: 5vh;
+        width: 5vh;
+      }
+    }
+  }
+
+  .articleProfilePhoto {
+    margin-left: 2vw;
+    height: 4vh;
+    border-radius: 70%;
+  }
+
+  .articlePhotoArea {
+    height: 40vh;
+    position: relative;
+    overflow: hidden;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    .articlePhoto {
+      max-height: 100%;
+      width: auto;
+    }
+  }
+
+  .articleTextContent {
+    padding: 1vh 2vw;
+    height: 10vh;
+    overflow: auto;
+  }
+
+  .articleBtnArea {
+    display: flex;
+    position: relative;
+    align-items: center;
+    margin-left: 2vw;
+    margin-right: 2vw;
+    margin-top: 2vh;
+    height: 3vh;
+
+    .articleLikeBtn {
+      padding: 0;
+      height: 3vh;
+      width: 3vh;
+      background-color: inherit;
+      border: 0;
+
+      img {
+        height: 3vh;
+        width: 3vh;
+      }
+    }
+
+    .articleReplyBtn,
+    .articleSaveBtn {
+      padding: 0;
+      margin-left: 2vw;
+      height: 3vh;
+      width: 3vh;
+      background-color: inherit;
+      border: 0;
+
+      img {
+        height: 3vh;
+        width: 3vh;
+      }
+    }
+
+    .articleBookmarkBtn {
+      position: absolute;
+      right: 0;
+      height: 3vh;
+      width: 3vh;
+      border: 0;
+      background-color: inherit;
+
+      img {
+        position: absolute;
+        top: 0;
+        right: 0;
+        height: 3vh;
+        width: 3vh;
+        border: 0;
+        background-color: inherit;
+      }
+    }
+  }
+
+  .articleLikeArea {
+    height: 4vh;
+    display: flex;
+    align-items: center;
+    margin-left: 2vw;
+    margin-right: 2vw;
+    margin-top: 1vh;
+    border-bottom: 1px solid rgb(188, 188, 188);
+
+    .whoFirstLikePhoto {
+      height: 3vh;
+      border-radius: 70%;
+    }
+
+    .infoLikeWho {
+      margin-left: 1vw;
+      height: 3vh;
+      width: 60vw;
+    }
+  }
+
+  .articleReplyArea {
+    background-color: inherit;
+    height: 11vh;
+    width: 68vw;
+    position: relative;
+    margin-left: 2vw;
+    display: flex;
+    overflow: auto;
+
+    .replyWriterArea {
+      li {
+        margin-top: 1vh;
+      }
+    }
+
+    .replyContentArea {
+      li {
+        margin-top: 1vh;
+        margin-left: 7vw;
+      }
+    }
+
+    .minibox1 {
+      height: 4vh;
+      width: 68vw;
+      display: flex;
+      align-items: center;
+
+      .miniboxPhoto1 {
+        height: 3vh;
+        border-radius: 70%;
+      }
+
+      .miniboxNick1 {
+        height: 4vh;
+        width: 15vw;
+        margin-left: 1vw;
+        display: flex;
+        align-items: center;
+        overflow: hidden;
+      }
+
+      .miniboxComment1 {
+        height: 4vh;
+        width: 40vw;
+        display: flex;
+        align-items: center;
+        overflow: hidden;
+      }
+
+      .deleteBtn1 {
+        border: 0;
+        background-color: inherit;
+        width: 3vh;
+        height: 3vh;
+        padding: 0;
+        position: absolute;
+        right: 8vw;
+
+        img {
+          border: 0;
+          position: absolute;
+          top: 0;
+          right: 0;
+          width: 3vh;
+          height: 3vh;
+        }
+      }
+
+      .showFullReplyModalBtn {
+        border: 0;
+        background-color: inherit;
+        color: grey;
+        position: absolute;
+        right: 2vw;
+      }
+    }
+
+    .minibox2 {
+      height: 4vh;
+      width: 68vw;
+      margin-top: 1vh;
+      display: flex;
+      align-items: center;
+
+      .miniboxPhoto2 {
+        height: 3vh;
+        border-radius: 70%;
+      }
+
+      .miniboxNick2 {
+        height: 4vh;
+        width: 15vw;
+        margin-left: 1vw;
+        display: flex;
+        align-items: center;
+        overflow: hidden;
+      }
+
+      .miniboxComment2 {
+        height: 4vh;
+        width: 40vw;
+        display: flex;
+        align-items: center;
+        overflow: hidden;
+      }
+
+      .deleteBtn2 {
+        border: 0;
+        background-color: inherit;
+        width: 3vh;
+        height: 3vh;
+        padding: 0;
+        position: absolute;
+        right: 8vw;
+
+        img {
+          border: 0;
+          position: absolute;
+          top: 0;
+          right: 0;
+          width: 3vh;
+          height: 3vh;
+        }
+      }
+    }
+
+    .articleAllReplyBtn {
+      position: absolute;
+      right: 0;
+      top: 0;
+      margin-right: 2vw;
+      border: 0;
+      padding: 0;
+      height: 3vh;
+      width: 4vw;
+      background-color: inherit;
+      color: grey;
+
+      span {
+        position: absolute;
+        right: 0;
+        top: 0;
+        width: 4vw;
+        height: 3vh;
+        padding: 0;
+      }
+    }
+  }
+
+  .articleInsertReplyArea {
+    background-color: inherit;
+    height: 4vh;
+    margin-left: 2vw;
+    position: relative;
+    display: flex;
+    align-items: flex-start;
+
+    .articleInsertReplyContent {
+      width: 60vw;
+      border: 0;
+
+      &:focus {
+        outline: none;
+      }
+    }
+
+    .articleInsertReplyBtn {
+      border: 0;
+      background-color: inherit;
+      color: rgb(147, 202, 208);
+      position: absolute;
+      right: 2vw;
+    }
+  }
+}
