+++ src/pages/jihyeon/Main/MainRight/MainRight.scss
@@ -1,0 +1,206 @@
+.mainRight {
+  width: 20vw;
+  margin-top: 10vh;
+  margin-left: 4vw;
+  margin-right: 2vw;
+  flex-direction: column;
+  background-color: inherit;
+
+  .sidebarLogined {
+    height: 11vh;
+    display: none;
+    align-items: flex-start;
+    flex-direction: column;
+    background-color: inherit;
+    position: relative;
+
+    .sidebarLoginedTop {
+      height: 5vh;
+      display: flex;
+      align-items: center;
+      position: relative;
+
+      img {
+        height: 5vh;
+        display: flex;
+        align-items: center;
+        border-radius: 70%;
+      }
+
+      span {
+        width: 18vw;
+        margin-left: 1vw;
+      }
+    }
+
+    .sidebarLogOutBtn {
+      background-color: white;
+      color: black;
+      border: 0;
+      width: 20vw;
+      height: 4vh;
+      position: absolute;
+      bottom: 1vh;
+      border: 1px solid rgb(179, 179, 179);
+    }
+  }
+
+  .sidebarNotLogined {
+    display: '';
+
+    button {
+      background-color: white;
+      color: black;
+      border: 0;
+      width: 20vw;
+      height: 4vh;
+    }
+
+    .sidebarJoinBtn {
+      margin-top: 1vh;
+    }
+  }
+
+  .sidebarStory {
+    height: 27vh;
+    width: 20vw;
+    display: flex;
+    align-items: flex-start;
+    flex-direction: column;
+    border: 1px solid rgb(179, 179, 179);
+    margin-top: 2vh;
+    padding-top: 1.5vh;
+    padding-left: 2vw;
+    padding-right: 2vw;
+    background-color: white;
+    position: relative;
+
+    .sidebarStoryTop {
+      width: 20vw;
+      display: flex;
+      align-items: flex-end;
+
+      span {
+        font-size: 1.5vh;
+        color: grey;
+      }
+
+      button {
+        background-color: inherit;
+        border: 0;
+        position: absolute;
+        right: 1vw;
+        font-weight: bold;
+        font-size: 1.5vh;
+      }
+    }
+
+    .sidebarStoryMain {
+      width: 15vw;
+      height: 23vh;
+
+      .story {
+        display: flex;
+        align-items: center;
+        height: 5.5vh;
+
+        img {
+          width: 3.5vh;
+          border-radius: 70%;
+        }
+      }
+
+      button {
+        border: 0;
+        background-color: inherit;
+        width: 18vw;
+        height: 5vh;
+        font-size: 2.5vh;
+        position: absolute;
+        left: 2vw;
+      }
+    }
+  }
+
+  .sidebarRecommend {
+    height: 27vh;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    background-color: white;
+    position: relative;
+    border: 1px solid rgb(179, 179, 179);
+    margin-top: 2vh;
+    padding-top: 1.5vh;
+    padding-left: 2vw;
+    padding-right: 2vw;
+
+    .sidebarRecommendTop {
+      width: 20vw;
+      display: flex;
+      align-items: flex-end;
+
+      span {
+        font-size: 1.5vh;
+        color: grey;
+      }
+
+      button {
+        background-color: inherit;
+        border: 0;
+        position: absolute;
+        right: 1vw;
+        font-weight: bold;
+        font-size: 1.5vh;
+      }
+    }
+
+    .sidebarRecommendMain {
+      width: 15vw;
+      height: 23vh;
+
+      .recommendWho {
+        display: flex;
+        align-items: center;
+        height: 5.5vh;
+
+        img {
+          width: 3.5vh;
+          border-radius: 70%;
+        }
+
+        button {
+          border: 0;
+          background-color: inherit;
+          width: 18vw;
+          height: 5vh;
+          font-size: 2.5vh;
+          position: absolute;
+          left: 2vw;
+        }
+      }
+    }
+  }
+
+  .sidebarFooter {
+    margin-top: 2vh;
+    display: flex;
+    color: grey;
+    flex-direction: column;
+    align-items: center;
+
+    button {
+      font-size: 2vh;
+      width: 20vw;
+      margin-top: 2vh;
+      border: 0;
+      background-color: inherit;
+      color: grey;
+    }
+
+    span {
+      margin-top: 2vh;
+      color: grey;
+    }
+  }
+}
