--- src/pages/jihyeon/Main/ReplyJihyeon/ReplyJihyeon.scss
@@ -1,18 +1,0 @@
-.replyJihyeon {
-  display: flex;
-
-  height: 3vh;
-  width: 68vw;
-
-  .replyNickNameArea {
-    margin-top: 1vh;
-    margin-left: 1vw;
-  }
-
-  .replyContentArea {
-    position: absolute;
-    left: 20vw;
-
-    margin-top: 1vh;
-  }
-}
