--- src/pages/jihyeon/Main/replyListJihyeon.js
@@ -1,28 +1,0 @@
-const REPLY_LIST_JIHYEON = [
-  {
-    id: 1,
-    userName: 'wecode',
-    content: 'Welcome to world best coding bootcamp!',
-    isLiked: true,
-  },
-  {
-    id: 2,
-    userName: 'joonsikyang',
-    content: 'Hi there.',
-    isLiked: false,
-  },
-  {
-    id: 3,
-    userName: 'jayPark',
-    content: 'Hey.',
-    isLiked: false,
-  },
-  {
-    id: 4,
-    userName: 'jayPark',
-    content: 'Hey.',
-    isLiked: false,
-  },
-];
-
-export default REPLY_LIST_JIHYEON;
